requirements.txt
asgiref==3.6.0
black==23.3.0
boto3==1.26.105
botocore==1.29.105
click==8.1.3
Django==4.1.7
django-admin==2.0.2
django-admin-material-dashboard==1.0.9
django-cors-headers==3.14.0
django-excel-response2==3.0.5
django-extensions==3.2.3
django-material-admin==1.8.6
django-simple-history==3.3.0
django-six==1.0.5
django-storages==1.13.2
djangorestframework==3.14.0
excel-base==1.0.4
isoweek==1.3.3
jmespath==1.0.1
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.0
pandas==2.1.3
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.18.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
s3transfer==0.6.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.3
TimeConvert==3.0.13
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
xlwt==1.3.0
